fix: core:data モジュールに data を移した
