--- app/src/main/java/app/doggy/mybrary/data/api/BaseUrl.kt
@@ -1,5 +1,0 @@
-package app.doggy.mybrary.data.api
-
-object BaseUrl {
-  const val BOOK_BASE_URL = "https://www.googleapis.com/books/v1/"
-}
