--- app/src/main/java/app/doggy/mybrary/data/db/Migration.kt
@@ -1,10 +1,0 @@
-package app.doggy.mybrary.data.db
-
-import androidx.room.migration.Migration
-import androidx.sqlite.db.SupportSQLiteDatabase
-
-val MIGRATION_1_2 = object : Migration(1, 2) {
-  override fun migrate(database: SupportSQLiteDatabase) {
-    database.execSQL("ALTER TABLE books ADD COLUMN books_api_id TEXT")
-  }
-}
