+++ core/data/src/main/java/app/doggy/mybrary/core/data/api/BaseUrl.kt
@@ -1,0 +1,5 @@
+package app.doggy.mybrary.core.data.api
+
+object BaseUrl {
+  const val BOOK_BASE_URL = "https://www.googleapis.com/books/v1/"
+}
