+++ core/data/src/main/java/app/doggy/mybrary/core/data/db/Migration.kt
@@ -1,0 +1,10 @@
+package app.doggy.mybrary.core.data.db
+
+import androidx.room.migration.Migration
+import androidx.sqlite.db.SupportSQLiteDatabase
+
+val MIGRATION_1_2 = object : Migration(1, 2) {
+  override fun migrate(database: SupportSQLiteDatabase) {
+    database.execSQL("ALTER TABLE books ADD COLUMN books_api_id TEXT")
+  }
+}
