+++ core/data/src/main/java/app/doggy/mybrary/core/data/db/dao/BookAuthorCrossRefDao.kt
@@ -1,0 +1,11 @@
+package app.doggy.mybrary.core.data.db.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import app.doggy.mybrary.core.data.db.entity.BookAuthorCrossRef
+
+@Dao
+interface BookAuthorCrossRefDao {
+  @Insert
+  fun insert(bookAuthorCrossRef: BookAuthorCrossRef): Long
+}
