+++ core/data/src/main/java/app/doggy/mybrary/core/data/db/di/DaoModule.kt
@@ -1,0 +1,20 @@
+package app.doggy.mybrary.core.data.db.di
+
+import app.doggy.mybrary.core.data.db.MybraryDatabase
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@InstallIn(SingletonComponent::class)
+@Module
+object DaoModule {
+  @Provides
+  fun provideBookDao(db: MybraryDatabase) = db.bookDao()
+
+  @Provides
+  fun provideDiaryDao(db: MybraryDatabase) = db.diaryDao()
+
+  @Provides
+  fun provideAuthorDao(db: MybraryDatabase) = db.authorDao()
+}
